Create replay event screen.
