+++ pages/events/TouchlessHomeInspections12May2020.vue
@@ -1,7 +1,7 @@
 <script>
 export default {
     middleware ({redirect }) {
-        return redirect('/events/TouchlessHomeInspectionsMay2020')
+        return redirect('/events/ReplayTouchlessHomeInspectionsMay2020')
     }
 }
 </script>
